Rename MissionSettings to MissionSettingsPanel to match name of React component
