+++ src/web/containers/MissionSettingsPanel/MissionSettingsPanel.less
